requirements.txt
aiosqlite==0.20.0
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
dill==0.3.9
ecdsa==0.19.0
greenlet==3.1.1
inquirerpy==0.3.4
Jinja2==3.1.5
MarkupSafe==2.1.5
multiprocess==0.70.14
nestd==0.3.1
orjson==3.9.15
passlib==1.7.4
pfzy==0.3.4
prompt_toolkit==3.0.48
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.0.1
python-jose==3.3.0
robyn==0.64.2
rsa==4.9
rustimport==1.3.4
six==1.17.0
SQLAlchemy==2.0.36
toml==0.10.2
typing_extensions==4.12.2
watchdog==4.0.1
wcwidth==0.2.13
